+++ argopy/plot/plot.py
@@ -37,7 +37,6 @@ if has_ipywidgets:
     import ipywidgets
 
 
-@warnUnless(has_ipython, "requires IPython to work as expected, only URLs are returned otherwise")
 def open_sat_altim_report(WMO: Union[str, list] = None, embed: Union[str, None] = "dropdown"):
     """ Insert the CLS Satellite Altimeter Report figure in notebook cell
 
@@ -62,6 +61,7 @@ def open_sat_altim_report(WMO: Union[str, list] = None, embed: Union[str, None] 
         Requires IPython to work as expected. If IPython is not available only URLs are returned.
 
     """
+    warnUnless(has_ipython, "requires IPython to work as expected, only URLs are returned otherwise")
     # Create the list of URLs and put them in a dictionary with WMO as keys:
     WMOs = check_wmo(WMO)
     urls = []
@@ -97,7 +97,6 @@ def open_sat_altim_report(WMO: Union[str, list] = None, embed: Union[str, None] 
         return urls_dict
 
 
-@warnUnless(has_mpl, "requires matplotlib installed")
 def plot_trajectory(
     df: pd.core.frame.DataFrame,
     style: str = STYLE["axes"],
@@ -146,6 +145,9 @@ def plot_trajectory(
     :class:`argopy.plot.scatter_map`.
 
     """
+
+    warnUnless(has_mpl, "requires matplotlib installed")
+
     with axes_style(style):
         # Set up the figure and axis:
         defaults = {"figsize": (10, 6), "dpi": 90}
@@ -220,7 +222,6 @@ def plot_trajectory(
     return fig, ax
 
 
-@warnUnless(has_mpl, "requires matplotlib installed")
 def bar_plot(
     df: pd.core.frame.DataFrame,
     by: str = "institution",
@@ -255,6 +256,9 @@ def bar_plot(
     fig: :class:`matplotlib.figure.Figure`
     ax: :class:`matplotlib.axes.Axes`
     """
+
+    warnUnless(has_mpl, "requires matplotlib installed")
+
     if by not in df:
         raise ValueError("'%s' is not a valid field for a bar plot" % by)
     with axes_style(style):
@@ -269,8 +273,6 @@ def bar_plot(
         ax.set_ylabel("")
     return fig, ax
 
-
-@warnUnless(has_mpl and has_cartopy, "requires matplotlib AND cartopy installed")
 def scatter_map(
         data: Union[xr.Dataset, pd.core.frame.DataFrame],
         x: Union[str] = None,
@@ -291,123 +293,7 @@ def scatter_map(
         set_global: bool = False,
         **kwargs
 ):
-    """Hello world !
-
-    Try-to-be generic function to create a scatter plot on a map from **argopy** :class:`xarray.Dataset` or :class:`pandas.DataFrame` data
-
-    Each point is an Argo profile location, colored with a user defined variable and colormap. Floats trajectory can be plotted or not.
-
-    Note that all parameters have default values.
-
-    Warnings
-    --------
-    This function requires `Cartopy <https://scitools.org.uk/cartopy/docs/latest/>`_.
-
-    Examples
-    --------
-    ::
-
-        from argopy.plot import scatter_map
-        from argopy import DataFetcher
-
-        ArgoSet = DataFetcher(mode='expert').float([6902771, 4903348]).load()
-        ds = ArgoSet.data.argo.point2profile()
-        df = ArgoSet.index
-
-        scatter_map(df)
-        scatter_map(ds)
-        scatter_map(ds, hue='DATA_MODE')
-        scatter_map(ds, hue='PSAL_QC')
-
-    ::
-
-        from argopy import OceanOPSDeployments
-        df = OceanOPSDeployments([-90, 0, 0, 90]).to_dataframe()
-        scatter_map(df, hue='status_code', traj=False)
-        scatter_map(df, x='lon', y='lat',
-                    hue='status_code',
-                    traj=False,
-                    cmap='deployment_status')
-
-    Parameters
-    ----------
-    data: :class:`xarray.Dataset` or :class:`pandas.DataFrame`
-        Input data structure
-    x: str, default=None
-        Name of the data variable to use as longitude.
-        If ``x`` is set to None, we'll try to guess which variable to use among standard names.
-    y: str, default=None
-        Name of the data variable to use as latitude.
-        If ``y`` is set to None, we'll try to guess which variable to use among standard names.
-    hue: str, default=None
-        Name of the data variable to use for points coloring.
-        If ``hue`` is set to None, we'll try to guess which variable to use to color points according to WMO.
-
-    Returns
-    -------
-    fig: :class:`matplotlib.figure.Figure`
-    ax: :class:`matplotlib.axes.Axes`
-
-    Other Parameters
-    ----------------
-    markersize: int, default=36
-        Size of the marker used for profiles location.
-    markeredgesize: float, default=0.5
-        Size of the marker edge used for profiles location.
-    markeredgecolor: str, default='default'
-        Color to use for the markers edge. The default color is 'DARKBLUE' from :class:`argopy.plot.ArgoColors.COLORS`
-
-    cmap: str, default=None
-        Colormap to use for points coloring. If set to None, we'll try to guess the most appropriate colormap for the
-        ``hue`` argument by matching it to values in :class:`argopy.plot.ArgoColors.list_valid_known_colormaps`.
-
-    traj: bool, default=True
-        Set to True in order to plot each float trajectories, i.e. join with a line all profiles from a single platform.
-    traj_axis: str, default='wmo'
-        Name of the data variable to use in order to determine profiles group making a single trajectory.
-    traj_color: str, default='default'
-        The unique color to use for all trajectories. The default color is the ``markeredgecolor`` value.
-
-    legend: bool, default=True
-        Display or not a legend for hue colors meaning. If the legend is too large, it can be removed with ``ax.get_legend().remove()``
-    legend_title: str, default='default'
-        String title of the legend box. By default, it is set to the ``hue`` value.
-    legend_location: str, default='upper right'
-        Location of the legend box. This is passed to the ``loc`` argument of :class:`~matplotlib:matplotlib.legend.Legend`.
-
-    set_global: bool, default=False
-        Force the map to be global.
-
-    kwargs
-        All other arguments are passed to :class:`matplotlib.figure.Figure.subplots`
-
-    """
-    pass
-
-@warnUnless(has_mpl and has_cartopy, "requires matplotlib AND cartopy installed")
-def scatter_map2(
-        data: Union[xr.Dataset, pd.core.frame.DataFrame],
-        x: Union[str] = None,
-        y: Union[str] = None,
-        hue: Union[str] = None,
-        markersize: int = 36,
-        markeredgesize: float = 0.5,
-        markeredgecolor: str = 'default',
-        cmap: Union[str] = None,
-        traj: bool = True,
-        traj_axis: Union[str] = None,
-        traj_color: str = 'default',
-        legend: bool = True,
-        legend_title: str = 'default',
-        legend_location: Union[str, int] = 0,
-        cbar: bool = False,
-        cbarlabels: Union[str, list] = 'auto',
-        set_global: bool = False,
-        **kwargs
-):
-    """Hello world !
-
-    Try-to-be generic function to create a scatter plot on a map from **argopy** :class:`xarray.Dataset` or :class:`pandas.DataFrame` data
+    """Try-to-be generic function to create a scatter plot on a map from **argopy** :class:`xarray.Dataset` or :class:`pandas.DataFrame` data
 
     Each point is an Argo profile location, colored with a user defined variable and colormap. Floats trajectory can be plotted or not.
 
@@ -493,6 +379,8 @@ def scatter_map2(
         All other arguments are passed to :class:`matplotlib.figure.Figure.subplots`
 
     """
+    warnUnless(has_mpl and False, "requires matplotlib AND cartopy installed")
+
     if isinstance(data, xr.Dataset) and data.argo._type == "point":
         # data = data.argo.point2profile(drop=True)
         raise InvalidDatasetStructure('Function only available to a collection of profiles')